gestdj-web-ui/requirements.txt
# Core dependencies for GesteDJ Web Backend
websockets>=11.0.0
numpy>=1.24.0
opencv-python>=4.8.0

# MediaPipe for hand detection
mediapipe>=0.10.0

# MIDI dependencies
python-rtmidi>=1.5.0
mido>=1.3.0

# Additional utilities
scipy>=1.10.0
matplotlib>=3.7.0

# Optional: For better performance
uvloop>=0.17.0; sys_platform != "win32"

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0